+++ CHANGELOG.md
@@ -17,6 +17,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Endpoint to clear Local AAC Cache
 - Script to create transfer bundles
+- CronEventEmitter, which emits fake Audit Log Events on a set interval
+- Observable Sensemaker, which processes all observables every 15 minutes
+  - Implemented geofence logic, other observable types to be implemented
 
 ### Changed
 - Updated Inference Sensemaker to only accept Observation Audit Log Events